Created SpotbotPlaylister
Adds scraper spotbot_playlister.py which will find the top recent songs
on a music subreddit and add them to a Spotify playlist. Also can create
brand new playlists based on the best songs of the past month or year.

Scripts have been added to ease the running of this daily job, as well
as seeding the subreddits we want to use.